Merge branch galaxy-collection-merge-main in to 'main' (partial).
This is a partial merge of:
  galaxy-collection-merge-main
into:
  main

Mostly to get the renamed role; which also happens to be the most
recent.  Ansible-galaxy (legacy) role import is limited (requires default
branch to have the structure of a role, and can't apparently be a
subfolder).  Further we want a single codebase for the role content.

Need to revisit architecture; probably turn this into a
galaxy-collection, i.e., namespace.collection.role, which allows
playbooks to be distributed with the content.

Should probably open an issue...